todo list
This project aims to help you manage your tasks more efficiently and in an organized way. Tasks are divided into three different groups and the user starts the process by clicking on the button of the group that he/she wants to add the task to. When you complete your task, you can instantly see how long it took to finish. In this way, you can track your work and optimize your time management.